//Resetted Attempt to include Images for Enum Case for Answer Structure
