+++ IntroductionViewControllerIMG.swift
@@ -1,0 +1,43 @@
+//
+//  ViewController.swift
+//  PersonalityQuiz
+//
+//  Created by Student04 on 2019-03-26.
+//  Copyright © 2019 Student04. All rights reserved.
+//
+
+import UIKit
+
+class IntroductionViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+    }
+    //Attempt To Assign Image to Enum Case:
+   /* @IBOutlet weak var globe: UIImageView!
+    @IBOutlet weak var bell: UIImageView!
+    @IBOutlet weak var book: UIImageView!
+    @IBOutlet weak var beaker: UIImageView!
+    
+        
+    enum ItemType {
+        case globeIcon
+        case bookIcon
+        case beakerIcon
+        case bellIcon
+        
+        var image: UIImage {
+            switch self {
+            case .globeIcon: return UIImage(named: "globe")! //optional value since UIImage?
+            case .bookIcon: return UIImage(named: "book")!
+            case .beakerIcon: return UIImage(named: "beaker")!
+            case .bellIcon: return UIImage(named: "bell")!
+
+            }
+        }
+    }
+   let globePersonalityResult: ItemType = .globeIcon
+*/
+    
+}
